Add WindowHandle::spawn
Window now accepts an executor on construction that will be used when
views want to schedule some work in the background. In xray_server we
are already supplying a concrete CpuPool executor, whereas in a
web-based version we could roll our own executor that's based on
`requestIdleCallback`.

A couple of things about the implementation:

* The executor is stored on the window as Box<future::Executor>. The
main reason for this design choice is so that the generic type of the
executor doesn't ripple through all the other types that reference
Window (e.g., views). This shouldn't be a big deal, as the cost of
invoking methods via the vtable should be negligible.
* The type of futures that the executor can spawn is also a Box. This is
because the executor trait is generic over the type of futures it can
spawn. This should only be temporary though, as the new version of the
future crate moves the generic future type into the `spawn` method.